Add files via upload
This Arduino code is designed to create a function for controlling 4 LDRs (Light Dependent Resistors) and 2 Servo Motors, allowing you to build a solar tracker. Additionally, the code includes a section that outputs the average values of the LDRs along with the corresponding time. To complement this functionality, there is an accompanying HTML code that works in conjunction with the Arduino code, enabling you to view the results in a web page.